Changes for x64 port.
git-svn-id: http://svn.code.sf.net/p/spimsimulator/code@422 663c0655-963b-4438-a250-40b5657b8e18

+++ PCSpim/PCSpim.cpp
@@ -227,7 +227,7 @@ void CPCSpimApp::RunMessageLoop()
 
 int CPCSpimApp::GetSetting(LPCTSTR strName, int nDefVal)
 {
-  return GetProfileInt("Settings", strName, nDefVal);
+  return (int)GetProfileInt("Settings", strName, nDefVal);
 }
